+++ sdk/core/azure-core/src/http/http.cpp
@@ -3,6 +3,8 @@
 
 #include "azure/core/http/http.hpp"
 
+#include <utility>
+
 void Azure::Core::Http::Details::InsertHeaderWithValidation(
     std::map<std::string, std::string>& headers,
     std::string const& headerName,
@@ -152,4 +154,14 @@ void Azure::Core::Http::Details::InsertHeaderWithValidation(
   }
   // insert (override if duplicated)
   headers[headerName] = headerValue;
+}
+
+std::string Azure::Core::Http::Internal::RangeToString(Range range)
+{
+  std::string rangeString = "bytes=" + std::to_string(range.Offset) + "-";
+  if (range.Length != std::numeric_limits<decltype(range.Length)>::max())
+  {
+    rangeString.append(std::to_string(range.Length));
+  }
+  return rangeString;
 }